nepali-legal-qa/backend/requirements.txt
# ── FastAPI server ────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.12

# ── SLM inference (HyDE generation) ──────────────────────────────
torch>=2.4.0
transformers==4.56.2
accelerate>=1.2.0

# ── LangChain stack (from nepali_rag_qa.ipynb) ───────────────────
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters

# ── Vector store ──────────────────────────────────────────────────
faiss-cpu

# ── Embeddings — sentence-transformers/LaBSE ─────────────────────
sentence-transformers>=3.3.0

# ── HuggingFace dataset loading (corpus download) ─────────────────
datasets>=3.1.0

# ── Utilities ─────────────────────────────────────────────────────
numpy>=1.26.0
pyngrok
python-dotenv>=1.0.0

# ── Authentication ─────────────────────────────────────────────
google-auth-oauthlib>=1.2.0
google-auth>=2.28.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.1